Refact: move v2 components to deprecated
