+++ components/common/Scroll/ScrollNavigator.tsx
@@ -2,7 +2,7 @@
 
 import classNames from "classnames/bind";
 import { useEffect } from "react";
-import { useNavigatorScroll } from "../../Home/scroll";
+import { useNavigatorScroll } from "../../deprecated/Home/scroll";
 import styles from "./ScrollNavigator.module.scss";
 const cx = classNames.bind(styles);
