+++ components/deprecated/Cover/Cover.module.scss
