+++ components/deprecated/Home/About/About.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import classNames from "classnames/bind";
-import useWaffleScroll from "../../../library/deprecated/waffleScroll";
+import useWaffleScroll from "../../../../library/deprecated/waffleScroll";
 import styles from "./About.module.scss";
 
 const cx = classNames.bind(styles);
