+++ components/deprecated/Home/Activity/Activity.module.scss
