+++ components/deprecated/Home/Billboard/Billboard.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useRef } from "react";
 import classNames from "classnames/bind";
-import useWaffleScroll from "../../../library/deprecated/waffleScroll";
+import useWaffleScroll from "../../../../library/deprecated/waffleScroll";
 import styles from "./Billboard.module.scss";
 import { dummyBanners } from "./dummy";
