+++ components/deprecated/Home/QnA/QnA.module.scss
