+++ components/deprecated/Home/QnA/QnA.tsx
@@ -1,9 +1,9 @@
 "use client";
 import classNames from "classnames/bind";
 import { useRef } from "react";
-import useDelayedState from "../../../hooks/delayedState/useDelayedState";
-import { useScroll } from "../../../hooks/scroll/useScroll";
-import useWaffleScroll from "../../../library/deprecated/waffleScroll";
+import useDelayedState from "../../../../hooks/delayedState/useDelayedState";
+import { useScroll } from "../../../../hooks/scroll/useScroll";
+import useWaffleScroll from "../../../../library/deprecated/waffleScroll";
 import styles from "./QnA.module.scss";
 
 const cx = classNames.bind(styles);
