+++ components/deprecated/Home/ServiceSummary/ServicesSummary.tsx
@@ -1,9 +1,9 @@
 "use client";
 
 import classNames from "classnames/bind";
-import { useScroll } from "../../../hooks/scroll/useScroll";
-import useDelayedState from "../../../hooks/delayedState/useDelayedState";
-import useWaffleScroll from "../../../library/deprecated/waffleScroll";
+import { useScroll } from "../../../../hooks/scroll/useScroll";
+import useDelayedState from "../../../../hooks/delayedState/useDelayedState";
+import useWaffleScroll from "../../../../library/deprecated/waffleScroll";
 import styles from "./ServicesSummary.module.scss";
 
 const cx = classNames.bind(styles);
