+++ components/deprecated/Home/WhatWeDo/WhatWeDo.tsx
@@ -1,8 +1,8 @@
 "use client";
 import classNames from "classnames/bind";
-import { useScroll } from "../../../hooks/scroll/useScroll";
-import useDelayedState from "../../../hooks/delayedState/useDelayedState";
-import useWaffleScroll from "../../../library/deprecated/waffleScroll";
+import { useScroll } from "../../../../hooks/scroll/useScroll";
+import useDelayedState from "../../../../hooks/delayedState/useDelayedState";
+import useWaffleScroll from "../../../../library/deprecated/waffleScroll";
 import styles from "./WhatWeDo.module.scss";
 
 const cx = classNames.bind(styles);
