+++ components/deprecated/Home/WhoAreWe/WhoAreWe.module.scss
