+++ components/deprecated/Home/scroll.ts
@@ -1,6 +1,6 @@
 "use client";
 
-import { createGlobalScrollHook } from "../../library/waffleScroll";
+import { createGlobalScrollHook } from "../../../library/waffleScroll";
 
 export const useNavigatorScroll = createGlobalScrollHook({
   globalState: { currentSection: "" },
