+++ components/newHome/About/About.tsx
@@ -2,7 +2,7 @@
 
 import classNames from "classnames/bind";
 import { useState } from "react";
-import { useNavigatorScroll } from "../../Home/scroll";
+import { useNavigatorScroll } from "../../deprecated/Home/scroll";
 import styles from "./About.module.scss";
 
 const cx = classNames.bind(styles);
