+++ components/newHome/Activity/Activity.tsx
@@ -2,7 +2,7 @@
 
 import classNames from "classnames/bind";
 import { MouseEvent, useState } from "react";
-import { useNavigatorScroll } from "../../Home/scroll";
+import { useNavigatorScroll } from "../../deprecated/Home/scroll";
 import styles from "./Activity.module.scss";
 
 const cx = classNames.bind(styles);
